third_party/WebKit/LayoutTests/platform/win/fast/text/font-features/caps-casemapping-expected.txt

This is a testharness.js-based test.
FAIL Synthetic small caps should produce the same width as manually uppercased, downscaled text. assert_equals: expected 3709.171875 but got 3501.5625
FAIL Strings containing uppercased sharp S should not be truncated. assert_equals: expected 70.03125 but got 57.515625
Harness: the test ran to completion.
